Refactor app cmds into files
